\name{pegas-package}
\alias{pegas-package}
\alias{pegas}
\docType{package}
\title{
Population and Evolutionary Genetics Analysis System
}
\description{
  \pkg{pegas} provides functions for the analysis of allelic data and of
  haplotype data from DNA sequences. It requires and complements two
  other R-packages: \pkg{ape} and \pkg{adegenet}.

  The complete list of functions can be displayed with
  \code{library(help = pegas)}.

  More information on \pkg{pegas} can be found at
  \url{http://ape-package.ird.fr/pegas.html}.
}
\author{
  Emmanuel Paradis, Alastair Potts, Klaus Schliep, David Winter

  Maintainer: Emmanuel Paradis
}
\keyword{package}
